优化了搜索文章列表，任何人都可以查看文章列表(可以通过 title ，tags ，content，并且 title 搜索出来高亮模式)

+++ model/model.go
@@ -8,7 +8,7 @@ type MODEL struct {
 	UpdateTime time.Time `gorm:"default:CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP" json:"-"`
 }
 type ParamDeleteList struct {
-	IDList []int64 `json:"id_list" binding:"required"`
+	IDList []uint `json:"id_list" binding:"required"`
 }
 type ParamList struct {
 	Page  int    `json:"page" form:"page"`
